requirements.txt
colorama
pyserial